requirements.txt
flask
redis
yfinance==0.2.10
psycopg2-binary
beautifulsoup4
requests
SQLAlchemy
pandas
selenium
